+++ quelware-core/python/src/quelware_core/entities/result/__init__.py
@@ -4,10 +4,14 @@ from typing import TypeAlias
 from quelware_core.entities.waveform.sampled import IqPoint, IqWaveform
 
 CaptureWindowName: TypeAlias = str
-IqResult: TypeAlias = list[IqWaveform] | list[IqPoint]
 
 
 @dataclass
 class ResultContainer:
-    iq_result: dict[CaptureWindowName, IqResult] = field(default_factory=dict)
+    iq_waveform_result: dict[CaptureWindowName, list[IqWaveform]] = field(
+        default_factory=dict
+    )
+    iq_point_result: dict[CaptureWindowName, list[IqPoint]] = field(
+        default_factory=dict
+    )
     integer_result: dict[CaptureWindowName, list[int]] = field(default_factory=dict)